Refactored ExitMenuTooltip
