+++ frontend/src/app/components/SideBar/UserInfoCard.js
@@ -2,7 +2,7 @@ import {Button, Card, Dropdown, Menu, Skeleton, Avatar } from "antd";
 // import Avatar from "antd/es/avatar/avatar";
 import {UpOutlined} from "@ant-design/icons";
 import React, {Component} from 'react';
-import ExitMenuTooltip from "./ExitMenuTooltip";
+import ExitMenuTooltip from "../ExitMenuTooltip";
 const { Meta } = Card;
 
 function setUserName(){
