+++ frontend/src/app/tests/ExitMenuTooltip.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import ExitMenuTooltip from "../components/SideBar/ExitMenuTooltip";
+import ExitMenuTooltip from "../components/ExitMenuTooltip";
 
 it('ExitMenuTooltip renders correctly', () => {
     const tree = renderer
